requirements.txt
flask==2.3.3
websocket-client==1.6.3
requests==2.31.0
gunicorn==21.2.0
